restaurant REST API Postman tests

+++ RestaurantProject/DataStore/RestaurantRepository.ts
@@ -1,4 +1,5 @@
 import {Schema, model, connect} from 'mongoose';
+import { transpileModule } from 'typescript';
 import Restaurant from '../CoreBusiness/RestaurantModel';
 
 export class RestaurantRepository
@@ -30,14 +31,6 @@ export class RestaurantRepository
                 email: 'someEmail@something.com',
                 website: 'someWebsite.com'
             },
-            {
-                name: 'Restaurant2',
-                address: 'Address2',
-                phone: '987654321',
-                nip: '987654321',
-                email: 'someEmail@somethingElse.com',
-                website: 'someOtherWebsite.com'
-            }
         ];
 
         if(await this.RestaurantModel.countDocuments() === 0)
@@ -54,7 +47,7 @@ export class RestaurantRepository
         }
     }
 
-    async addRestaurant(restaurant: Restaurant) : Promise<void>
+    async addRestaurant(restaurant: Restaurant) : Promise<boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
 
@@ -62,25 +55,44 @@ export class RestaurantRepository
         .create(restaurant)
         .then(function()
         {
-            console.log("Restaurant " + restaurant.name + " has been added!");
+            console.log('Restaurant ' + restaurant.name + ' has been added!');
         }).catch(function(err)
         {
             console.log(err);
         });
+
+        // check if restaurant has been added
+        const result = await this.RestaurantModel.findOne({name: restaurant.name});
+        
+        if (result)
+            return true;
+        else
+            return false;
+        
     }
 
-    async deleteRestaurantByName(restaurantName: string) : Promise<void>
+    async deleteRestaurantByName(restaurantName: string) : Promise<boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
+
+        const exists = await this.RestaurantModel.exists({name: restaurantName});
+        if (!exists)
+        {
+            console.log('Restaurant ' + restaurantName + ' does not exist!');
+            return false;
+        }
+
         await this.RestaurantModel
         .deleteOne({name: restaurantName})
         .then(function()
         {
-            console.log("Restaurant " + restaurantName + " has been deleted!")
+            console.log("Restaurant " + restaurantName + " has been deleted!");
         }).catch(function(err)
         {
             console.log(err);
         });
+
+        return true;
     }
     
     async getRestaurantByName(restaurantName: string) : Promise<Restaurant>
@@ -99,7 +111,7 @@ export class RestaurantRepository
         return await this.RestaurantModel.find({});
     }
 
-    async updateRestaurant(restaurantName: string, restaurant: Restaurant) : Promise<void>
+    async updateRestaurant(restaurantName: string, restaurant: Restaurant) : Promise<boolean>
     {
         await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
 
@@ -130,19 +142,13 @@ export class RestaurantRepository
             {
                 console.log(err);
             });
+            return true;
         }
         else
+        {
             console.log("Restaurant " + restaurantName + " does not exist!");
-    }
-
-    async CheckIfRestaurantExists(restaurantName: string) : Promise<boolean>
-    {
-        await connect('mongodb+srv://username:username123@cluster.itsrg.mongodb.net/RestaurantDb?retryWrites=true&w=majority');
-        let restaurant = await this.RestaurantModel.findOne({name: restaurantName});
-        if (restaurant)
-            return true;
-        else
             return false;
+        }
     }
 }
